modif route delete pour securité

+++ routes/affiches.js
@@ -28,30 +28,30 @@ router.get("/", async (req, res) => {
 // Post an affiche
 
 router.post("/", upload.single('file'), async (req, res) => {
-    try {
-      const resultCloudinary = await cloudinary.uploader.upload(req.file.path, {
-        folder: "Affiches",
-      });
-  
-      const newAffiche = new Affiche({
-        imageName: resultCloudinary.secure_url,
-        idCloud: resultCloudinary.public_id,
-        filmName: req.body.filmName,
-        realName: req.body.realName,
-        creationDate: new Date()
-      });
-  
-      const affiche = await newAffiche.save();
-      await unlinkAsync(req.file.path); // Assurez-vous d'effacer le fichier temporaire
-      res.json({ result: true, affiche });
-  
-    } catch (error) {
-      console.error('An error occurred:', error);
-      await unlinkAsync(req.file.path); // En cas d'erreur, effacez aussi le fichier
-      res.status(500).json({ result: false, error: error.message });
-    }
+  try {
+    const resultCloudinary = await cloudinary.uploader.upload(req.file.path, {
+      folder: "Affiches",
+    });
+
+    const newAffiche = new Affiche({
+      imageName: resultCloudinary.secure_url,
+      idCloud: resultCloudinary.public_id,
+      filmName: req.body.filmName,
+      realName: req.body.realName,
+      creationDate: new Date()
+    });
+
+    const affiche = await newAffiche.save();
+    await unlinkAsync(req.file.path); // Assurez-vous d'effacer le fichier temporaire
+    res.json({ result: true, affiche });
+
+  } catch (error) {
+    console.error('An error occurred:', error);
+    await unlinkAsync(req.file.path); // En cas d'erreur, effacez aussi le fichier
+    res.status(500).json({ result: false, error: error.message });
+  }
 });
-  
+
 
 // delete une affiche
 router.delete("/:id", async (req, res) => {
@@ -59,15 +59,23 @@ router.delete("/:id", async (req, res) => {
     const afficheId = req.params.id;
     // Trouver l'affiche dans la base de données
     const affiche = await Affiche.findById(afficheId);
+
     if (!affiche) {
       return res.status(404).json({ result: false, message: "Affiche not found" });
     }
 
     // Supprimer l'image de Cloudinary
-    await cloudinary.uploader.destroy(affiche.idCloud);
+    const result = await await cloudinary.uploader.destroy(affiche.idCloud);
 
-    // Supprimer l'affiche de la base de données MongoDB
-    await Affiche.deleteOne({ _id: afficheId });
+    console.log(result)
+    
+    if (result === 'ok') {
+      // Supprimer l'affiche de la base de données MongoDB
+      await Affiche.deleteOne({ _id: afficheId });
+    } else {
+      res.json({ result: false, message: "Document not found" });
+    }
+
 
     res.json({ result: true, message: "Affiche deleted successfully" });
   } catch (error) {
